+++ examples/opencode/plugins/xmpp-bridge.js
@@ -47,7 +47,7 @@
  */
 
 export const XmppBridgePlugin = async ({ client, directory, $ }) => {
-  const PLUGIN_VERSION = "0.8.14"
+  const PLUGIN_VERSION = "0.8.15"
   const pluginRef = (() => {
     try {
       // eslint-disable-next-line no-undef